fix(x64): add pointer-sized integer types, fix truncating casts
Windows is LLP64, so `long` stays 32-bit on x86-64 and cannot hold a
pointer. Adds `UnsignedIntPtr`/`IntPtr` (uintptr_t/intptr_t) to
BaseTypeCore.h and applies them to the 18 measured pointer-to-integer
truncation sites across Core (registry.cpp, Except.cpp, debug_debug.cpp,
debug_stack.cpp, huffencode.cpp, and the WWAudio Miles/callback chain).

Wire vs. runtime: `uint32` (bittype.h) is left alone everywhere -- it is
the on-disk/network/savegame wire format and must stay 4 bytes regardless
of target. `MILES_HANDLE` (AudibleSound.h) is widened to UnsignedIntPtr
because it is runtime-only, holds live Miles Sound System pointers, and is
defined in our own code rather than the fetched Miles stub.

Two of the "cast" sites were real x64 defects, not cast noise:
Except.cpp's two GetProcAddress-fill loops and debug_stack.cpp's gDbg
union both walk a `long unsigned int` stride across consecutive
function-pointer slots that are 8 bytes wide on Win64. Left as `long`,
the stride would silently corrupt every other function pointer at
runtime. Fixed the holder/stride, not just the visible cast -- this is
the strongest argument in this diff for why -fpermissive stays banned:
it would have hidden a genuine memory-corruption bug behind a warning.

SoundScene.cpp's EVENT_LOGICAL_HEARD dispatch smuggles two pointers
through `On_Event`'s uint32 params. Traced the full chain before
touching anything: the real declaration/definition is
SoundSceneObj.h's SoundSceneObjClass::On_Event (not AudioEvents.h, which
holds an unrelated and unused-for-this-path callback typedef family),
has no overrides, and only one call site passes real pointers. Widened
just that method's two params and the one call site; confirmed the
change did not need to spread into AudioEvents.h or the engine.

registry.cpp stored an HKEY in an `int` (also removes a stale
`assert(sizeof(HKEY) == sizeof(int))` that was one line away from
tripping on any 64-bit debug build once Key's type changed).

huffencode.cpp needed intptr_t via `<Utility/stdint_adapter.h>`, not a
raw `<stdint.h>` -- VC6 predates C99 and doesn't have the latter. The
shim is the same one BaseTypeCore.h already uses and was already on this
target's include path via corei_always -> core_utility.

Gates:
- G1 (32-bit control): exit 0, 0 errors, 31,555 warnings -- unchanged
  from baseline. (The registry.cpp fix does not remove any 32-bit
  warnings: HKEY<->int is a no-op reinterpret at that width, so GCC
  never warned there. The ~17 -Wint-to-pointer-cast warnings this fix
  does remove exist only in the x64 build log.)
- G2 (x64): errors 82->60, distinct shapes 41->29, affected files 10->5,
  `loses precision` occurrences 37->15, build ceiling 15%->100%.
- G3 (VC6 retail): exit 0, 0 errors, all seven .text digests
  byte-identical to docs/x64/baseline-vc6.md. .rdata/.data movement in
  the larger executables traced to resources/gitinfo/gitinfo.cpp.in
  build metadata (commit SHA, dates, timestamps) -- zero code
  differences.

+++ Core/Libraries/Source/Compression/EAC/huffencode.cpp
@@ -22,6 +22,7 @@
 #define __HUFWRITE 1
 
 #include <string.h>
+#include <Utility/stdint_adapter.h>
 #include "codex.h"
 #include "huffcodex.h"
 
@@ -1050,8 +1051,8 @@ static void HUFF_pack(struct HuffEncodeContext *EC,
 		if (!i3)
 			HUFF_writecode(EC,dest,i);
 
-		if (((long) bptr1- (long) EC->buffer) >= (long)(EC->plen+curpc))
-			curpc = (long) bptr1 - (long) EC->buffer - EC->plen;
+		if (((intptr_t) bptr1- (intptr_t) EC->buffer) >= (intptr_t)(EC->plen+curpc))
+			curpc = (intptr_t) bptr1 - (intptr_t) EC->buffer - EC->plen;
 	}
 
 	/* write EOF ([clue] 0gn [10]) */
